moved some functions, login ready?
